Address review: use react.block and move isSmartBlockUid
- Replace useEffect+renderBlock with roamAlphaAPI.ui.components.react.block
- Simplify isSmartBlockUid to only check :SmartBlock: text
- Move isSmartBlockUid to getLeftSidebarSettings (shared by both consumers)
- Restore hasSmartBlockSyntax as local fn in createDiscourseNode
